Get an image path from a clock

Get the path (or a different example path) to an image for a specific clock.

Usage

find_image_path(time, folder, clock, path = NULL, github = TRUE)

Arguments

time

character string in hms format. Which time does the image show?

folder

character string showing the clock specific folder format (output from list_times)

clock

character string. For which clock is additional information required?

path

character string. Path to the Github repo heike/ticktock-images (or a local clone).

github

binary value; true if the give path points to a github repository

Value

path to an image showing the specified time (based on its name, not the clock hand). In case of a warning the image corresponding to the closest time is shown.

Examples

# example code

find_image_path("18:21:55", "clock2019-10-16_18", "omni", "heike/ticktock-images", TRUE)
#> [1] "omni/clock2019-10-16_18/clock2019-10-16_18.21.55.jpg"
find_image_path("18:59:59", "clock2019-10-16_18", "omni", "heike/ticktock-images", TRUE)
#> Warning: Image for time  '18H 59M 59S' not found (in first 1000 images). Closest time is shown, it's 1661 seconds away. 
#> Repo uses format:
#> clock2019-10-16_18.14.42.jpg 
#> clock2019-10-16_18.14.43.jpg 
#> clock2019-10-16_18.14.44.jpg 
#> clock2019-10-16_18.14.45.jpg 
#> clock2019-10-16_18.14.46.jpg
#> [1] "omni/clock2019-10-16_18/clock2019-10-16_18.32.18.jpg"