Get a list of all available times for a clock

Using the output from list_clocks, find the available times for a clock.

Usage

list_times(clock, path = NULL, github = TRUE)

Arguments

clock

character string. For which clock is additional information required?

path

character string. Path to the Github repo heike/ticktock-images (or a local clone).

github

binary value; true if the give path points to a github repository

Value

vector of available clock recordings

Examples

# example code
list_times("omni", "heike/ticktock-images", TRUE)
#> [1] "clock2019-10-16_18"
list_times(list_clocks("heike/ticktock-images"))
#> [1] "clock2019-10-16_18"